refactored some common methods
